desktop-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biomotion-desktop"
version = "0.1.0"
description = "Desktop OpenSim ground-truth pipeline + iOS-export comparator for the BioMotion app."
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "BioMotion" }]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
plot = ["matplotlib>=3.7"]
test = ["pytest>=7.4"]

[project.scripts]
biomotion-desktop = "biomotion_desktop.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
